refactor(hub): move cache → kernel/
