+++ packages/hub/src/collection-config.ts
@@ -48,7 +48,7 @@ import type { CollectionMeta } from './with-shape/introspection/meta.js'
 import type { RefDescriptor } from './refs.js'
 import type { JoinableSource } from './query/index.js'
 import { VectorSet, type EmbeddingDescriptor } from './with-lookup/embeddings/index.js'
-import { Lru } from './cache/index.js'
+import { Lru } from './kernel/cache/index.js'
 import type { ReadOnlyVaultFacade } from './with-audit/guards/types.js'
 import type { DerivationRegistry } from './with-formula/derivations/registry.js'
 import type { TxContext } from './with-commit/tx/transaction.js'
