+++ packages/hub/src/collection.ts
@@ -86,7 +86,7 @@ import type { RefDescriptor } from './refs.js'
 import { buildDescription, deriveZodFields, type CollectionDescription, type DescribeOptions } from './with-shape/introspection/describe.js'
 import { buildJsonSchema } from './with-shape/introspection/json-schema.js'
 import type { CollectionConfig } from './with-shape/introspection/types.js'
-import { Lru, parseBytes, estimateRecordBytes, type LruStats } from './cache/index.js'
+import { Lru, parseBytes, estimateRecordBytes, type LruStats } from './kernel/cache/index.js'
 import { generateULID } from './with-pod/ulid.js'
 import type { PresenceHandle, PresenceHandleOpts } from './with-party/team/presence.js'
 import type { SyncStrategy } from './with-party/team/sync-strategy.js'
