+++ packages/hub/src/kernel/enclave/record-keys/record-codec.ts
@@ -26,7 +26,7 @@ import { DebugReservedFieldError } from '../../../errors.js'
 import { validateSchemaOutput, type StandardSchemaV1 } from '../../../schema.js'
 import type { CrdtMode, CrdtState } from '../../../with-commit/crdt/crdt.js'
 import type { CrdtStrategy } from '../../../with-commit/crdt/strategy.js'
-import type { Lru } from '../../../cache/index.js'
+import type { Lru } from '../../cache/index.js'
 
 /** Everything the moving crypto methods touched on `this.*`, as a flat context. */
 export interface RecordCodecContext<T> {
